test(arcor2_arserver): server tests divided into multiple modules
- shared fixtures located in conftest.py
- `start_processes` now checks whether ARServer started or died
- ports for services are picked-up randomly so tests can run in parallel
