+++ src/python/arcor2_arserver/tests/objects/BUILD
@@ -1,0 +1,1 @@
+python_library()
